fix: use platform modifier key (Ctrl on Windows) and f32 font sizes
M7: Change default keybindings from Cmd+ to Ctrl+ on non-macOS platforms.
    Applied to both config.rs defaults and keybindings.rs with_defaults().
M8: Change font_size fields from u32 to f32 in AppearanceSettings and
    TerminalSettings. Removes all as-f32 casts at use sites.

+++ crates/codirigent-ui/src/workspace/impl_settings.rs
@@ -48,9 +48,9 @@ impl WorkspaceView {
             };
 
             let theme = self.workspace.theme();
-            user_settings.appearance.font_size = theme.font_size_base as u32;
+            user_settings.appearance.font_size = theme.font_size_base;
             user_settings.appearance.grid_gap = theme.grid_gap as u32;
-            user_settings.terminal.font_size = theme.terminal_font_size as u32;
+            user_settings.terminal.font_size = theme.terminal_font_size;
 
             // Use pre-cached results from background detection (started at init).
             // Falls back to synchronous detection if background task hasn't finished yet.
